pyproject.toml
[tool.poetry]
name = "physical-app"
version = "0.1.0"
description = ""
authors = ["Nikolaev Valentin <val8673@gmail.com>"]
readme = "README.md"
packages = [{include = "physical_app"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
python-dotenv = "^0.21.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
psycopg2 = "^2.9.5"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
drf-yasg = "^1.21.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
